30/01/2018 17:30

+++ migrations/m171121_000001_keys_menu.php
@@ -13,22 +13,59 @@ class m171121_000001_keys_menu extends Migration {
      */
     public function safeUp() {
 
-        $pos = $this->db->createCommand('SELECT id FROM menu WHERE name = "pos" AND parent IS NULL AND route IS NULL')->queryOne();
+        $posS = $this->db->createCommand('SELECT id FROM menu WHERE name = "pos" AND parent IS NULL AND route IS NULL')->queryOne();
+        /*         * ************************************* */
+        /* Удаляем старые записи */
+        $auth = new rbac\DbManager();
+        $pos = $auth->getPermission('Pos');
 
-        $this->batchInsert('menu', ['parent', 'name', 'route', 'order'], [
-            [$pos['id'], 'Работа с ключами', '/pos/keys', 100],
-        ]);
+        if (($hkcv = $auth->getPermission('/pos/key-reserve')) != false) {
+            $auth->removeChild($pos, $hkcv);
+            $auth->remove($hkcv);
+        }
 
+        if (($hkcv = $auth->getPermission('/pos/enter-key-reserve')) != false) {
+            $auth->removeChild($pos, $hkcv);
+            $auth->remove($hkcv);
+        }
+
+        if (($posS) != false) {
+            $this->delete('menu', [
+                'route' => '/pos/key-reserve',
+                'parent' => $posS['id']
+            ]);
+            $this->delete('menu', [
+                'route' => '/pos/enter-key-reserve',
+                'parent' => $posS['id']
+            ]);
+        }
+        /*         * ********************************* */
+
+        if (($posS) != false) {
+            $this->batchInsert('menu', ['parent', 'name', 'route', 'order'], [
+                [$posS['id'], 'Работа с ключами', '/pos/keys', 100],
+                [$posS['id'], 'Зарезервировать транспортный ключ', '/pos/key-reserve/O03D2', 20],
+                [$posS['id'], 'Зарезервировать рабочий ключ', '/pos/key-reserve/O03S2', 21],
+            ]);
+        }
         /*
          * добавляем права доступа
          */
         $route = new rbac\Item();
         $route->type = rbac\Item::TYPE_PERMISSION;
-        $route->name = '/pos/keys';
 
         $auth = new rbac\DbManager();
         $perm = $auth->getPermission('Pos');
 
+        $route->name = '/pos/keys';
+        $auth->add($route);
+        $auth->addChild($perm, $route);
+
+        $route->name = '/pos/key-reserve/O03D2';
+        $auth->add($route);
+        $auth->addChild($perm, $route);
+
+        $route->name = '/pos/key-reserve/O03S2';
         $auth->add($route);
         $auth->addChild($perm, $route);
     }
@@ -44,11 +81,12 @@ class m171121_000001_keys_menu extends Migration {
         $auth->removeChild($pos, $hkcv);
         $auth->remove($hkcv);
 
-        $pos = $this->db->createCommand('SELECT id FROM menu WHERE name = "pos" AND parent IS NULL AND route IS NULL')->queryOne();
-        $this->delete('menu', [
-            'route' => '/pos/keys',
-            'parent' => $pos['id']
-        ]);
+        if (($pos = $this->db->createCommand('SELECT id FROM menu WHERE name = "pos" AND parent IS NULL AND route IS NULL')->queryOne()) != false) {
+            $this->delete('menu', [
+                'route' => '/pos/keys',
+                'parent' => $pos['id']
+            ]);
+        }
         #return false;
     }
